flightTelemetry_persist/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.6
requests>=2.32
python-dateutil>=2.9
